requirements.txt
altair==5.3.0
annotated-types==0.6.0
arrow==1.3.0
asgiref==3.7.2
astroid==2.6.2
attrs==23.2.0
binaryornot==0.4.4
black==22.12.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.3
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
cleo==2.1.0
click==8.0.1
colorama==0.4.6
cookiecutter==2.4.0
copier==9.1.1
crashtest==0.4.1
crispy-bootstrap5==0.7
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.0.8
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.1.3
django-filter==23.5
django-jet-reboot==1.3.5
django-model-utils==4.3.1
djangorestframework==3.14.0
dulwich==0.21.7
dunamai==1.19.2
fastjsonschema==2.19.1
filelock==3.12.4
Flask==2.0.1
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
googletrans==3.1.0a0
gopay==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.22.2
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.1
installer==0.7.0
isort==5.9.2
itsdangerous==2.0.1
jaraco.classes==3.3.1
Jinja2==3.1.3
jinja2-ansible-filters==1.3.2
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.0
lazy-object-proxy==1.6.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mccabe==0.6.1
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.7
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.2
pathspec==0.11.2
pexpect==4.9.0
Pillow==9.3.0
pipenv==2023.10.20
pkginfo==1.9.6
platformdirs==3.11.0
plumbum==1.8.2
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.16.1
pylint==2.9.3
PyPDF2==3.0.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-slugify==8.0.1
pytube==10.9.2
pytz==2022.7
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyyaml-include==1.3.2
questionary==2.0.1
rapidfuzz==3.6.1
referencing==0.34.0
regex==2024.4.16
reportlab==4.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.6.0
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.4.4
streamlit==1.33.0
stripe==5.5.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.4.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
transformers==4.40.1
trove-classifiers==2024.1.31
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
virtualenv==20.24.6
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==2.0.1
whitenoise==6.2.0
wrapt==1.12.1
youtube-dl==2021.12.17
zipp==3.17.0
